backend/requirements.txt
asgiref==3.8.1
boto3==1.35.2
botocore==1.35.2
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
Django==4.2.3
django-admin-autocomplete-filter==0.7.1
django-admin-list-filter-dropdown==1.0.3
django-admin-shortcuts==2.1.1
django-cors-headers==4.1.0
django-extensions==3.2.3
django-filter==23.2
django-json-widget==1.1.1
django-oauth-toolkit==2.3.0
django-password-reset==2.0
django-queryable-properties==1.9.3
django-ses==3.5.0
django-shell-plus==1.1.7
djangoql==0.17.1
djangorestframework==3.14.0
djangorestframework-csv==2.1.1
future==1.0.0
ipython==8.15.0
idna==3.7
jmespath==1.0.1
jwcrypto==1.5.6
jwt==1.3.1
oauthlib==3.2.2
ply==3.11
psycopg2==2.9.7
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
unicodecsv==0.14.1
urllib3==2.2.2